Add files via upload
This is a consolidate list of the common and additional networks protocols mentioned during the Course 3- Modulo 2: Networks operations. It included the main information such as acronyms, name, short explanation, description and ports used (if applied). It can facilitate studying this topic that looks challenge the 1st time that you read